Rename open pull request templates, and settings from opened to open
This is in preparation of adding support for closed pull requests.  By
changing the config setting, and template file names to be open,
they'll match the GitHub API end-point, and we can simply key off of
that internally everywhere, when determining which config setting, and
templates to use when sending notifications.