Add support for notifying on closed pull requests
We now respect the alert_on_close setting to send out notifications
when there are closed pull requests for a repository.  This also
supports grouping these notifications by the repository they are for.

We still do not keep track of which pull requests we've sent
notifications for, so once a pull request has been closed, there will
be a notification sent about it on every run.