Clean-up the grammar in the sample closed templates
The original wording was awkward, and misleading.  "[T]he" didn't
belong in the original wording given the sentence structure, and is
better phrased as "On {{closed_at}} pull request #{{number}} was
closed.".

The additional information in the partial was also misleading, since
the user in question wasn't the one that closed the pull request, but
was the original requester instead.