Tighten up regex used to validate repository configuration keys
Previously, we weren't anchoring the regex used to validate the
repository keys, which meant that bad data would validate as long as
there was a piece of it that looked like it was good.  Anchoring the
regex on both ends fixes this problem.

Also, we weren't allowing valid repository names that had a '.' in
them.  This is no longer a problem since we've added '.' to the
character class used to check the repository name portion of the
configuration key.