Add ability to specify just a GitHub account in the config file.
By allowing users to specify just the GitHub account name in the
config file instead of always requiring a repository specified by the
'<user-name>/<repository-name>' format, we will be able to support
"discovering" the repositories for that account.

Being able to discover the repositories that a user owns, will mean
that we will not require them to update the config file whenever they
add or remove a repository.